feat: checkout products

+++ back-end/src/database/migrations/20230328201911-sales.js
@@ -20,7 +20,7 @@ const sale = (Sequelize) => ({
   },
   userId: { type: Sequelize.INTEGER, ...dataUser, field: 'user_id' },
   sellerId: { type: Sequelize.INTEGER, ...dataSeller, field: 'seller_id' },
-  totalPrice: { type: Sequelize.FLOAT, allowNull: false, field: 'total_price' },
+  totalPrice: { type: Sequelize.DECIMAL(4, 2), allowNull: false, field: 'total_price' },
   deliveryAddress: { type: Sequelize.STRING, allowNull: false, field: 'delivery_address' },
   deliveryNumber: { type: Sequelize.STRING, allowNull: false, field: 'delivery_number' },
   saleDate: { type: Sequelize.DATE, allowNull: false, field: 'sale_date' },
